moved to totalassets
