+++ packages/appstore/src/components/total-assets/total-assets.tsx
@@ -1,20 +1,106 @@
 import React from 'react';
 import { Popover, Text } from '@deriv/components';
 import { localize } from '@deriv/translations';
-import { formatMoney } from '@deriv/shared';
+import { formatMoney, isDemo } from '@deriv/shared';
 import { TAccountCategory } from 'Types';
+import { DetailsOfEachMT5Loginid, Mt5LoginList } from '@deriv/api-types';
+import { observer } from 'mobx-react-lite';
+import { useStores } from 'Stores';
 
 type TTotalAssets = {
-    amount: number;
-    currency: string;
     category: TAccountCategory;
 };
 
-const TotalAssets = ({ amount, currency, category }: TTotalAssets) => {
+const TotalAssets = ({ category }: TTotalAssets) => {
+    const { client, common } = useStores();
+    const { obj_total_balance, dxtrade_accounts_list, mt5_login_list, accounts, account_list } = client;
+    const { getExchangeRate } = common;
+
+    const [exchanged_rate_cfd_real, setExchangedRateCfdReal] = React.useState(1);
+    const [exchanged_rate_demo, setExchangedRateDemo] = React.useState(1);
+    const [exchanged_rate_cfd_demo, setExchangedRateCfdDemo] = React.useState(1);
+
+    const cfd_real_currency =
+        mt5_login_list.find((mt5_accounts: Mt5LoginList) => !isDemo(mt5_accounts))?.currency ||
+        dxtrade_accounts_list.find((mt5_accounts: Mt5LoginList) => !isDemo(mt5_accounts))?.currency;
+
+    const cfd_demo_currency =
+        mt5_login_list.find((mt5_accounts: Mt5LoginList) => isDemo(mt5_accounts))?.currency ||
+        dxtrade_accounts_list.find((mt5_accounts: Mt5LoginList) => isDemo(mt5_accounts))?.currency;
+
+    React.useEffect(() => {
+        const getCurrentExchangeRate = (
+            currency: string,
+            setExchangeRate: React.Dispatch<React.SetStateAction<number>>
+        ) => {
+            getExchangeRate(currency, account_total_balance_currency).then((res: number) => {
+                setExchangeRate(res);
+            });
+        };
+        if (cfd_real_currency !== account_total_balance_currency) {
+            getCurrentExchangeRate(cfd_real_currency, setExchangedRateCfdReal);
+        }
+        if (vrtc_currency !== account_total_balance_currency) {
+            getCurrentExchangeRate(vrtc_currency, setExchangedRateDemo);
+        }
+        if (cfd_demo_currency !== account_total_balance_currency) {
+            getCurrentExchangeRate(cfd_demo_currency, setExchangedRateCfdDemo);
+        }
+    }, []);
+
+    const vrtc_loginid = account_list.find((account: { is_virtual: boolean }) => account.is_virtual).loginid;
+    const vrtc_currency = accounts[vrtc_loginid] ? accounts[vrtc_loginid].currency : 'USD';
+    const account_total_balance_currency = obj_total_balance.currency;
+
+    const getTotalBalanceCfd = (mt5_accounts: Mt5LoginList, is_demo: boolean, exchange_rate: number) => {
+        return mt5_accounts
+            .filter((mt5_account: DetailsOfEachMT5Loginid) => (is_demo ? isDemo(mt5_account) : !isDemo(mt5_account)))
+            .reduce(
+                (
+                    total: {
+                        balance: number;
+                    },
+                    mt5_account: DetailsOfEachMT5Loginid
+                ) => {
+                    total.balance += (mt5_account?.balance ?? 1) * exchange_rate;
+                    return total;
+                },
+                { balance: 0 }
+            );
+    };
+
+    const getTotalDemoAssets = (): number => {
+        const vrtc_balance = accounts[vrtc_loginid] ? accounts[vrtc_loginid].balance : 0;
+        const mt5_demo_total = getTotalBalanceCfd(mt5_login_list, true, exchanged_rate_cfd_demo);
+        const dxtrade_demo_total = getTotalBalanceCfd(dxtrade_accounts_list, true, exchanged_rate_cfd_demo);
+
+        const total =
+            (vrtc_currency !== account_total_balance_currency ? vrtc_balance * exchanged_rate_demo : vrtc_balance) +
+            mt5_demo_total.balance +
+            dxtrade_demo_total.balance;
+
+        return total;
+    };
+
+    const getTotalRealAssets = (): number => {
+        const mt5_total = getTotalBalanceCfd(mt5_login_list, false, exchanged_rate_cfd_real);
+        const dxtrade_total = getTotalBalanceCfd(dxtrade_accounts_list, false, exchanged_rate_cfd_real);
+
+        let total = obj_total_balance.amount_real;
+
+        total += obj_total_balance.amount_mt5 > 0 ? obj_total_balance.amount_mt5 : mt5_total.balance;
+        total += obj_total_balance.amount_dxtrade > 0 ? obj_total_balance.amount_dxtrade : dxtrade_total.balance;
+
+        return total;
+    };
+
+    const currency = account_total_balance_currency;
+    const total_assets = category === 'real' ? getTotalRealAssets() : getTotalDemoAssets();
+
     return (
         <div className='total-assets'>
             <Text weight='bold' size='m' className='total-assets-amount'>
-                {formatMoney(currency, amount, true)}
+                {formatMoney(currency, total_assets, true)}
             </Text>
             <Text weight='bold' size='m' color='prominent' className='total-assets-currency'>
                 {currency}
@@ -35,4 +121,4 @@ const TotalAssets = ({ amount, currency, category }: TTotalAssets) => {
     );
 };
 
-export default TotalAssets;
+export default observer(TotalAssets);
