--- packages/appstore/src/modules/trading-hub/index.scss
@@ -1,10 +1,0 @@
-.trading-hub {
-    &__header {
-        &--right {
-            display: flex;
-            justify-content: flex-end;
-            align-items: center;
-            padding-top: 2rem;
-        }
-    }
-}
